feat: Partial Enity Rendering Impl ---
This is a quick update for the old code so it mostly *works* for the newer systems.

It's not pretty and still has issues but it renders an entity.

The /scripts folder in resources should go into run/renders.

Note: That this isn't updated to the latest dev system (1.16-merged-official Uses Official Mappings) or the 1.16-merged as I don't have time to split/port for fabric at the moment

Vendors the JOML Library (MIT License) so I don't have to deal with Mojang's impl.

Also not sure if this was the most up-to-date code but was the latest I could find.

Tagging @unascribed and @Tslat as you both were interested in this code.

Signed-off-by: AterAnimAvis <AterAnimAvis@gmail.com>